// backend/src/utils/types.ts
export type ApiResponse<T> = {
	success: boolean;
	message: string;
	data?: T;
	errors?: Record<string, unknown>;
};
// ...
export type KYCType = z.infer<typeof KYCSchema> & {
	id?: string;
	status: string;
	customer: {
		firstname: string;
		lastname: string;
		residentialAddress: string;
		phoneNo: string;
		email: string;
		dob: string;
		isVerified: boolean;
	};
};
export type KYCReportType = {
	dashdata: {
		customerReport: {
			totalCustomers: number;
			lastMonthTotalSignupCustomers: number;
			prevMonthTotalCustomers: number;
			growthPercentage: number;
			growthStatus: string;
		};
		totalApproved: number;
		totalReviewed: number;
		approvalRate: string;
	};
	kycs: KYCType[];
};